requirements.txt
Pillow==8.3.2
torch==1.9.1.post3
torchvision==0.10.0a0